dashboard/dev-requirements.txt
aiocache==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
aiofile==3.8.1 ; python_version >= "3.11" and python_version < "4"
aiohttp==3.8.4 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
alembic==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.11" and python_version < "4.0"
apscheduler==3.10.1 ; python_version >= "3.11" and python_version < "4.0"
async-lru==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.11" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.11" and python_version < "4.0"
caio==0.9.11 ; python_version >= "3.11" and python_version < "4"
certifi==2022.12.7 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.11" and python_version < "4.0"
cognitojwt[async]==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
coverage[toml]==7.2.3 ; python_version >= "3.11" and python_version < "4.0"
cryptography==39.0.1 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
factory-boy==3.2.1 ; python_version >= "3.11" and python_version < "4.0"
faker==18.4.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi-limiter==0.1.5 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.92.0 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.11" and python_version < "4.0"
graphql-core==3.2.3 ; python_version >= "3.11" and python_version < "4"
greenlet==2.0.2 ; python_version >= "3.11" and python_version < "4.0" and platform_machine == "aarch64" or python_version >= "3.11" and python_version < "4.0" and platform_machine == "ppc64le" or python_version >= "3.11" and python_version < "4.0" and platform_machine == "x86_64" or python_version >= "3.11" and python_version < "4.0" and platform_machine == "amd64" or python_version >= "3.11" and python_version < "4.0" and platform_machine == "AMD64" or python_version >= "3.11" and python_version < "4.0" and platform_machine == "win32" or python_version >= "3.11" and python_version < "4.0" and platform_machine == "WIN32"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
httpx-auth==0.15.0 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.23.3 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
libcst==0.4.9 ; python_version >= "3.11" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.11" and python_version < "4.0"
pydantic[email]==1.10.7 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.11" and python_version < "4.0"
pytest-asyncio==0.21.0 ; python_version >= "3.11" and python_version < "4.0"
pytest-cov==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytest==7.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.11" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.11" and python_version < "4.0"
python-slugify==8.0.1 ; python_version >= "3.11" and python_version < "4.0"
pytz-deprecation-shim==0.1.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.11" and python_version < "4.0"
redis==4.5.1 ; python_version >= "3.11" and python_version < "4.0"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
rich==13.3.1 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
sentry-sdk==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==67.6.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.4 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.11" and python_version < "4.0"
strawberry-graphql[debug-server]==0.158.2 ; python_version >= "3.11" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspect==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.11" and python_version < "4.0"
tzlocal==4.3 ; python_version >= "3.11" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn[standard]==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "4.0"
watchfiles==0.18.1 ; python_version >= "3.11" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.11" and python_version < "4.0"
